Commit Message:
```sh
chore(dependencies): remove @alpinejs/intersect module

This commit removes the "@alpinejs/intersect" dependency from both package.json and package-lock.json files. This could be due to it being unused, deprecated or replaced with a different solution. Ensure that this removal doesn't break any functionality which relied on this module.
```