feat(ui): add LogoElement and LogoGrid components, update index page
This commit adds two new reusable components: the LogoElement and LogoGrid. The LogoElement represents a logo with a hover state that displays the logo's name. The LogoGrid is a section that renders a grid of LogoElements.

The index page ("src/pages/index.astro") has been updated to incorporate these new elements. A new "Our Tech Stack" section was added which utilizes these components to display a number of tech logos. Image URL paths on this page have also been updated.